Add the disconnecting socket state
According the socket.io specs the the possible states for the socket
are `disconnected`, `disconnecting`, `connecting`, `connected`.
In this case when the `[AZSocketIO disconnect]` is called the socket
goes in the state AZSocketIOStateDisconnecting while asking to the
transport to effectively close the connection, after that the socket
will be in the AZSocketIOStateDisconnected state.